rework way we look for links and open them, refactor into separate util fund
